feat(cli): consolidate 'nono policy' subcommands under 'nono profile' with deprecation alias (#594)
Move the five nono policy subcommands — groups, profiles, show, diff,
validate — into the nono profile namespace as list, show, diff,
validate, and groups. Output, flags, exit codes, and JSON shapes are
preserved byte-for-byte; stdout now uses the unified 'nono profile:'
label.

The old 'nono policy <sub>' form continues to work via a dedicated
shim at crates/nono-cli/src/deprecated_policy.rs: every invocation
prints a one-line deprecation warning to stderr and delegates to the
canonical handler. The shim re-exports arg types via 'pub use' aliases
so there are no parallel type definitions to drift. Top-level
'nono policy --help' is labelled [deprecated]. Removal steps for when
the deprecation window closes are recorded in the shim's header
comment.

Includes new parser and integration tests for every migrated command,
a byte-grep unit test on the embedded authoring guide to prevent
future drift, a renamed integration-test file, refreshed site docs
(policy-introspection.mdx -> profile-introspection.mdx), a CHANGELOG
entry, and a rustfmt pass.
