+++ crates/nono-cli/src/main.rs
@@ -15,6 +15,7 @@ mod command_blocking_deprecation;
 mod command_runtime;
 mod config;
 mod credential_runtime;
+mod deprecated_policy;
 mod exec_strategy;
 mod execution_runtime;
 mod hooks;
@@ -28,7 +29,6 @@ mod output;
 mod package;
 mod package_cmd;
 mod policy;
-mod policy_cmd;
 mod profile;
 mod profile_cmd;
 mod profile_runtime;
