+++ docs/cli/internals/wsl2-feature-matrix.mdx
@@ -27,8 +27,7 @@ Legend: **Full** = identical to native Linux | **Blocked (default)** = fails sec
 | `nono logs` | Full | Session event logs, pure userspace |
 | `nono inspect` | Full | Session state, pure userspace |
 | `nono prune` | Full | Session cleanup, pure userspace |
-| `nono policy` (groups/profiles/show/diff/validate) | Full | Pure logic |
-| `nono profile` (init/schema/guide) | Full | Pure logic |
+| `nono profile` (init/list/show/diff/validate/groups/schema/guide) | Full | Pure logic |
 
 ## Filesystem Permissions
